add "Acellerator: None" on Creating a Workbench form
If the cluster has GPUs available, the Create a Workbench form will
display the options in an "Accelerator" dropdown in the "Deployment
size" section of the form.

Guide users to not select an Accelerator for this workshop.

Thanks Erwan Granger <32554862+erwangranger@users.noreply.github.com>
for pointing this out to me.

Co-authored-by: Melissa Flinn <31925014+MelissaFlinn@users.noreply.github.com>