add "Acellerator: None" on Creating a Workbench form (#98)
* add "Acellerator: None" on Creating a Workbench form

If the cluster has GPUs available, the Create a Workbench form will
display the options in an "Accelerator" dropdown in the "Deployment
size" section of the form.

Guide users to not select an Accelerator for this workshop.

Thanks Erwan Granger <32554862+erwangranger@users.noreply.github.com>
for pointing this out to me.

Co-authored-by: Melissa Flinn <31925014+MelissaFlinn@users.noreply.github.com>

* Update creating-a-workbench.adoc

made a small change (changed the line from a NOTE to a step) - since it is a required step if the Accelerators section appears, and also to avoid breaking the numbering of steps. (You'd need to add a + to  indent the NOTE under a step and to avoid restarting the step numbering)

---------

Co-authored-by: Melissa Flinn <31925014+MelissaFlinn@users.noreply.github.com>